Update deploy script

+++ deploy.sh
@@ -2,15 +2,18 @@
 
 set -e
 
-echo Uploading template
+echo Uploading template and assets
 aws s3 sync template/ s3://cloud.jamesoff.net/template --delete --exact-timestamps
+aws s3 sync inc/ s3://cloud.jamesoff.net/inc --delete --exact-timestamps
 
-echo Preparing lambda
-[ -f lambda.zip ] && rm lambda.zip
-(
-cd lambda
-zip -r ../lambda.zip ./*
-)
+if [[ $1 == "lambda" ]]; then
+	echo Preparing lambda
+	[ -f lambda.zip ] && rm lambda.zip
+	(
+	cd lambda
+	zip -q -r ../lambda.zip ./*
+	)
 
-echo Deploying lambda
-aws lambda update-function-code --function-name cloud-handler --zip-file fileb://lambda.zip
+	echo Deploying lambda
+	aws lambda update-function-code --function-name cloud-handler --zip-file fileb://lambda.zip
+fi
